requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
albumentations
numpy
matplotlib
pillow
opencv-python
scikit-learn
scikit-image
nibabel
gdown